requirements-dev.txt
-r requirements.txt
pytest
pre-commit
black
isort
mypy
types-requests
pandas-stubs
openai